feature: remove slug | change upload file from s3 amazon to static

+++ src/controllers/userController.js
@@ -4,10 +4,9 @@ const express = require("express");
 const fs = require("fs");
 const sharp = require('sharp');
 const bcrypt = require("bcrypt");
-const {PutObjectCommand, DeleteObjectCommand} = require("@aws-sdk/client-s3");
 
 // middlewares
-const {upload, client} = require("../middlewares/upload.js");
+const {upload} = require("../middlewares/upload.js");
 const {requireAuth} = require("../middlewares/authentication");
 
 // models
@@ -41,44 +40,31 @@ router.put("/editPassword", requireAuth, async (req, res) => {
 
 router.put("/editProfile", [requireAuth, upload.single("avatar")], async (req, res) => {
     try {
-        const {name, family, phoneNumber} = req.body;
+        const {name, family, phoneNumber, preview} = req.body;
 
         const user = await User.findById(res.locals.user.id);
 
-        let avatarPath = user.avatar;
+        let avatarPath = preview;
 
         if (req.file) {
-            const fileName = `avatar-compressed-${req.file.filename.replace(path.extname(req.file.filename), ".webp")}`;
-            const filePath = path.resolve("uploads", fileName);
 
-            await sharp(req.file.path)
-                .toFormat("webp")
-                .resize({width: 120, height: 120, fit: "cover"})
-                .toFile(filePath);
-
-            if (user.avatar) {
-                const fileName = path.basename(user.avatar);
-
-                const params = {
-                    Bucket: process.env.BUCKET_NAME,
-                    Key: fileName,
-                };
-
-                await client.send(new DeleteObjectCommand(params));
+            if (avatarPath) {
+                const fileName = path.basename(avatarPath);
+                const filePath = path.resolve("uploads" , "avatar" , fileName);
+                await fs.unlinkSync(filePath);
             }
 
-            avatarPath = path.join(process.env.ASSETS_URL, fileName);
+            const fileName = `${res.locals.user.id}-${req.file.filename}`;
+            const oldFilePath = req.file.path;
+            const newFilePath = path.resolve("uploads" , "avatar", fileName);
 
-            const params = {
-                Body: fs.readFileSync(filePath),
-                Bucket: process.env.BUCKET_NAME,
-                Key: fileName,
-            };
+            await sharp(oldFilePath)
+                .resize({width: 240, height: 240, fit: "cover"})
+                .toFile(newFilePath);
 
-            await client.send(new PutObjectCommand(params));
+            await fs.unlinkSync(oldFilePath);
 
-            await fs.unlinkSync(req.file.path);
-            await fs.unlinkSync(filePath);
+            avatarPath = process.env.ASSET_URL + "/avatar/" + fileName;
         }
 
         await User.findOneAndUpdate(
@@ -101,7 +87,6 @@ router.put("/editProfile", [requireAuth, upload.single("avatar")], async (req, r
 
         res.status(200).json({data: privateUser, message: "پروفایل اصلاح شد", status: "success"});
     } catch (err) {
-        console.log(err)
         res.status(500).json({message: "مشکلی در سرور به وجود آمده است", status: "failure"});
     }
 });
